+++ curifactory/utils.py
@@ -406,7 +406,6 @@ def args_hash(args, registry_path: str, store_in_registry: bool = False) -> str:
         with open(registry_path, "w") as outfile:
             json.dump(registry, outfile, indent=4, default=stringify)
 
-    print("regular hash called", hash_str)
     return hash_str
 
 
@@ -458,7 +457,6 @@ def add_args_combo_hash(
         registry[hash_str] = {"active": active_key, "arg_list": hashes}
         with open(registry_path, "w") as outfile:
             json.dump(registry, outfile, indent=4, default=lambda x: str(x))
-    print("combo hash called", hash_str)
     return hash_str